Save the state of bookmark when extension is closed
The user input is saved in localstorage when the extension popup is
closed. This does not mean the bookmark is created or added to the
database. It is only saved temporarily on the browser.
When the extension popup is opened again on the same URL, the saved
input from localstorage is used to fill the appropriate input boxes.
This does not happpen if the bookmark was saved, updated or deleted.